Move cors attribute to own file
