+++ src/Teapot.Web/Controllers/StatusController.cs
@@ -46,16 +46,4 @@ namespace Teapot.Web.Controllers
             return new CustomHttpStatusCodeResult((int)HttpStatusCode.OK, statusData);
         }
     }
-
-    public class CorsAllowAnyOriginAttribute : ActionFilterAttribute
-    {
-        public override void OnResultExecuted(ResultExecutedContext filterContext)
-        {
-            if (filterContext.HttpContext.Response.Headers["Access-Control-Allow-Origin"] == null)
-            {
-                filterContext.HttpContext.Response.Headers.Add("Access-Control-Allow-Origin", "*");
-            }
-            base.OnResultExecuted(filterContext);
-        }
-    }
 }
