+++ src/Teapot.Web/CorsAllowAnyOriginAttribute.cs
@@ -1,0 +1,16 @@
+using System.Web.Mvc;
+
+namespace Teapot.Web
+{
+    public class CorsAllowAnyOriginAttribute : ActionFilterAttribute
+    {
+        public override void OnResultExecuted(ResultExecutedContext filterContext)
+        {
+            if (filterContext.HttpContext.Response.Headers["Access-Control-Allow-Origin"] == null)
+            {
+                filterContext.HttpContext.Response.Headers.Add("Access-Control-Allow-Origin", "*");
+            }
+            base.OnResultExecuted(filterContext);
+        }
+    }
+}
